pyproject.toml
[tool.poetry]
name = "fklcproject"
version = "0.1.0"
description = ""
authors = ["shkoh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.40.8"
fastapi = "^0.112.1"
tiktoken = "^0.7.0"
sqlalchemy = "^2.0.32"
uvicorn = "^0.30.6"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
transformers = "^4.44.0"
pymysql = "^1.1.1"
aiomysql = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
